# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UOMI API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register",
                "description": "This endpoint is used to register a new user.",
                "operationId": "6cd4a8afb20be55c12cf820331f54ce4",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "User name",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Doe",
                                        "nullable": false
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "demo@example.com",
                                        "nullable": false
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "password",
                                        "nullable": false
                                    },
                                    "device_name": {
                                        "description": "The name of the device",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John's iPhone",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad Request - One or more errors with the input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "This endpoint is used to login a user. <br><br>You have to provide the following informations: <br>Email <br>Password<br>Device Name: This can be a name of the phone, or a browser name<br><br>The endpoint will return an user object, with a token, the token is a bearer token, which must be attached in the header of all request, this token is used to identify the user, in the system.",
                "operationId": "b12e417f73d4e905041e3e87641dca4d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "example": "demo@example.com",
                                        "nullable": false
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "example": "password",
                                        "nullable": false
                                    },
                                    "device_name": {
                                        "description": "The name of the device",
                                        "type": "string",
                                        "example": "John's iPhone",
                                        "nullable": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Group response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid credentials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Bad Request - One or more errors with the input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "BadRequestError": {
                "title": "Bad Request",
                "description": "Bad Request - The request was invalid or cannot be served. The exact error is sent in the message and further ellaborated for each section in the errors datapoint. E.g. „The JSON is not valid“",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "The errors returned when the request is invalid, formatted as an object with the key being the field name and the value being an array of errors.",
                        "type": "object",
                        "example": {
                            "field1": [
                                "The field1 is required.",
                                "The field1 must be at least 10 characters."
                            ],
                            "field2": [
                                "The field2 field is required."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "User": {
                "title": "User schema",
                "properties": {
                    "id": {
                        "description": "User id",
                        "type": "integer",
                        "example": 1,
                        "nullable": false
                    },
                    "name": {
                        "description": "User name",
                        "type": "string",
                        "example": "John Doe",
                        "nullable": false
                    },
                    "email": {
                        "description": "User email",
                        "type": "string",
                        "example": "demo@example.com",
                        "nullable": false
                    },
                    "email_verified_at": {
                        "description": "User email verified at",
                        "type": "string",
                        "default": null,
                        "example": "2021-01-01 00:00:00",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "User created at",
                        "type": "string",
                        "example": "2021-01-01 00:00:00",
                        "nullable": false
                    },
                    "updated_at": {
                        "description": "User updated at",
                        "type": "string",
                        "example": "2021-01-01 00:00:00",
                        "nullable": false
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}